Formações/ChatGPT for Devs/Criando um Podcast com IAs Generativas/requirements.txt
openai==2.21.0
requests==2.32.5
google-api-python-client==2.190.0
google-auth-oauthlib==1.2.4
python-dotenv==1.2.1